requirements.txt
discord.py==2.4.0
python-dotenv==1.0.1
streamlit==1.39.0
watchdog==5.0.3
